examples/push-ingestion/bigquery/push-iceberg-tables/requirements.txt
google-cloud-bigquery>=3.0.0
pycarlo>=0.12.251
python-dateutil>=2.8.0
